+++ examples/example_test.go
@@ -1,6 +1,7 @@
 package examples
 
 import (
+	"context"
 	"log"
 	"testing"
 
@@ -22,13 +23,16 @@ func (l Logger) Debug(stmt *sqlstmt.Statement) {
 
 // TestExamples :
 func TestExamples(t *testing.T) {
-	client, err := sqlike.Connect("mysql",
+	client, err := sqlike.Connect(
+		context.Background(),
+		"mysql",
 		options.Connect().
 			SetUsername("root").
 			SetPassword("abcd1234").
 			SetCharset("utf8mb4"),
 	)
 	require.NoError(t, err)
+	defer client.Close()
 
 	mg := connectMongoDB()