+++ sqlike/connection.go
@@ -1,6 +1,7 @@
 package sqlike
 
 import (
+	"context"
 	"database/sql"
 	"errors"
 	"log"
@@ -10,7 +11,7 @@ import (
 )
 
 // Open : connect to sql server with connection string
-func Open(driver string, opt *options.ConnectOptions) (client *Client, err error) {
+func Open(ctx context.Context, driver string, opt *options.ConnectOptions) (client *Client, err error) {
 	if opt == nil {
 		return nil, errors.New("sqlike: invalid connection options <nil>")
 	}
@@ -22,13 +23,13 @@ func Open(driver string, opt *options.ConnectOptions) (client *Client, err error
 	if err != nil {
 		return
 	}
-	client, err = newClient(driver, conn, dialect, opt.Charset, opt.Collate)
+	client, err = newClient(ctx, driver, conn, dialect, opt.Charset, opt.Collate)
 	return
 }
 
 // MustOpen :
-func MustOpen(driver string, opt *options.ConnectOptions) *Client {
-	client, err := Open(driver, opt)
+func MustOpen(ctx context.Context, driver string, opt *options.ConnectOptions) *Client {
+	client, err := Open(ctx, driver, opt)
 	if err != nil {
 		panic(err)
 	}
@@ -36,12 +37,12 @@ func MustOpen(driver string, opt *options.ConnectOptions) *Client {
 }
 
 // Connect :
-func Connect(driver string, opt *options.ConnectOptions) (client *Client, err error) {
-	client, err = Open(driver, opt)
+func Connect(ctx context.Context, driver string, opt *options.ConnectOptions) (client *Client, err error) {
+	client, err = Open(ctx, driver, opt)
 	if err != nil {
 		return
 	}
-	err = client.Ping()
+	err = client.PingContext(ctx)
 	if err != nil {
 		client.Close()
 		return
@@ -50,8 +51,8 @@ func Connect(driver string, opt *options.ConnectOptions) (client *Client, err er
 }
 
 // MustConnect will panic if cannot connect to sql server
-func MustConnect(driver string, opt *options.ConnectOptions) *Client {
-	conn, err := Connect(driver, opt)
+func MustConnect(ctx context.Context, driver string, opt *options.ConnectOptions) *Client {
+	conn, err := Connect(ctx, driver, opt)
 	if err != nil {
 		panic(err)
 	}